@@ -1,6 +1,6 @@
 name: shadcn_ui
 description: shadcn-ui ported in Flutter. Awesome UI components for Flutter, fully customizable.
-version: 0.4.3
+version: 0.5.5
 homepage: https://mariuti.com/shadcn-ui
 repository: https://github.com/nank1ro/flutter-shadcn-ui
 documentation: https://mariuti.com/shadcn-ui
@@ -12,17 +12,15 @@ environment:
   flutter: ">=1.17.0"
 
 dependencies:
-  boxy: ^2.2.0
   flutter:
     sdk: flutter
   flutter_animate: ^4.5.0
   flutter_localizations:
     sdk: flutter
   flutter_svg: ^2.0.10+1
-  lucide_icons_flutter: ^1.0.8
-  rive: ^0.13.1
+  lucide_icons_flutter: ^1.1.0
   # used by Table
-  two_dimensional_scrollables: ^0.2.1
+  two_dimensional_scrollables: ^0.3.1
   vector_graphics: ^1.1.11+1
 
 dev_dependencies:
