@@ -4,8 +4,63 @@ import 'package:awesome_flutter_extensions/awesome_flutter_extensions.dart';
 import 'package:example/common/base_scaffold.dart';
 import 'package:example/common/properties/bool_property.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:shadcn_ui/shadcn_ui.dart';
+
+const fruits = {
+  'apple': 'Apple',
+  'banana': 'Banana',
+  'blueberry': 'Blueberry',
+  'grapes': 'Grapes',
+  'pineapple': 'Pineapple',
+};
+
+const timezones = {
+  'North America': {
+    'est': 'Eastern Standard Time (EST)',
+    'cst': 'Central Standard Time (CST)',
+    'mst': 'Mountain Standard Time (MST)',
+    'pst': 'Pacific Standard Time (PST)',
+    'akst': 'Alaska Standard Time (AKST)',
+    'hst': 'Hawaii Standard Time (HST)',
+  },
+  'Europe & Africa': {
+    'gmt': 'Greenwich Mean Time (GMT)',
+    'cet': 'Central European Time (CET)',
+    'eet': 'Eastern European Time (EET)',
+    'west': 'Western European Summer Time (WEST)',
+    'cat': 'Central Africa Time (CAT)',
+    'eat': 'Eastern Africa Time (EAT)',
+  },
+  'Asia': {
+    'msk': 'Moscow Time (MSK)',
+    'ist': 'India Standard Time (IST)',
+    'cst_china': 'China Standard Time (CST)',
+    'jst': 'Japan Standard Time (JST)',
+    'kst': 'Korea Standard Time (KST)',
+    'ist_indonasia': 'Indonesia Standard Time (IST)',
+  },
+  'Australia & Pacific': {
+    'awst': 'Australian Western Standard Time (AWST)',
+    'acst': 'Australian Central Standard Time (ACST)',
+    'aest': 'Australian Eastern Standard Time (AEST)',
+    'nzst': 'New Zealand Standard Time (NZST)',
+    'fjt': 'Fiji Time (FJT)',
+  },
+  'South America': {
+    'art': 'Argentina Time (ART)',
+    'bot': 'Bolivia Time (BOT)',
+    'brt': 'Brasilia Time (BRT)',
+    'clt': 'Chile Standard Time (CLT)',
+  },
+};
+
+const frameworks = {
+  'nextjs': 'Next.js',
+  'svelte': 'SvelteKit',
+  'nuxtjs': 'Nuxt.js',
+  'remix': 'Remix',
+  'astro': 'Astro',
+};
 
 class SelectPage extends StatefulWidget {
   const SelectPage({super.key});
@@ -16,55 +71,14 @@ class SelectPage extends StatefulWidget {
 
 class _SelectPageState extends State<SelectPage> {
   bool enabled = true;
-  final focusNodes = [FocusNode(), FocusNode()];
+  final focusNodes = [FocusNode(), FocusNode(), FocusNode()];
+  var searchValue = '';
 
-  final fruits = {
-    'apple': 'Apple',
-    'banana': 'Banana',
-    'blueberry': 'Blueberry',
-    'grapes': 'Grapes',
-    'pineapple': 'Pineapple',
-  };
-
-  final timezones = {
-    'North America': {
-      'est': 'Eastern Standard Time (EST)',
-      'cst': 'Central Standard Time (CST)',
-      'mst': 'Mountain Standard Time (MST)',
-      'pst': 'Pacific Standard Time (PST)',
-      'akst': 'Alaska Standard Time (AKST)',
-      'hst': 'Hawaii Standard Time (HST)',
-    },
-    'Europe & Africa': {
-      'gmt': 'Greenwich Mean Time (GMT)',
-      'cet': 'Central European Time (CET)',
-      'eet': 'Eastern European Time (EET)',
-      'west': 'Western European Summer Time (WEST)',
-      'cat': 'Central Africa Time (CAT)',
-      'eat': 'Eastern Africa Time (EAT)',
-    },
-    'Asia': {
-      'msk': 'Moscow Time (MSK)',
-      'ist': 'India Standard Time (IST)',
-      'cst_china': 'China Standard Time (CST)',
-      'jst': 'Japan Standard Time (JST)',
-      'kst': 'Korea Standard Time (KST)',
-      'ist_indonasia': 'Indonesia Standard Time (IST)',
-    },
-    'Australia & Pacific': {
-      'awst': 'Australian Western Standard Time (AWST)',
-      'acst': 'Australian Central Standard Time (ACST)',
-      'aest': 'Australian Eastern Standard Time (AEST)',
-      'nzst': 'New Zealand Standard Time (NZST)',
-      'fjt': 'Fiji Time (FJT)',
-    },
-    'South America': {
-      'art': 'Argentina Time (ART)',
-      'bot': 'Bolivia Time (BOT)',
-      'brt': 'Brasilia Time (BRT)',
-      'clt': 'Chile Standard Time (CLT)',
-    },
-  };
+  Map<String, String> get filteredFrameworks => {
+        for (final framework in frameworks.entries)
+          if (framework.value.toLowerCase().contains(searchValue.toLowerCase()))
+            framework.key: framework.value
+      };
 
   @override
   void dispose() {
@@ -99,6 +113,14 @@ class _SelectPageState extends State<SelectPage> {
           onChanged: (value) => setState(
             () =>
                 value ? focusNodes[1].requestFocus() : focusNodes[1].unfocus(),
+          ),
+        ),
+        MyBoolProperty(
+          label: 'Framework Focused',
+          value: focusNodes[2].hasFocus,
+          onChanged: (value) => setState(
+            () =>
+                value ? focusNodes[2].requestFocus() : focusNodes[2].unfocus(),
           ),
         ),
       ],
@@ -148,8 +170,12 @@ class _SelectPageState extends State<SelectPage> {
                     textAlign: TextAlign.start,
                   ),
                 ),
-                ...zone.value.entries
-                    .map((e) => ShadOption(value: e.key, child: Text(e.value)))
+                ...zone.value.entries.map(
+                  (e) => ShadOption(
+                    value: e.key,
+                    child: Text(e.value),
+                  ),
+                )
               ],
             ),
           ),
@@ -159,6 +185,35 @@ class _SelectPageState extends State<SelectPage> {
                 .value[value];
             return Text(timezone!);
           },
+        ),
+        ShadSelect<String>.withSearch(
+          enabled: enabled,
+          focusNode: focusNodes[2],
+          minWidth: 180,
+          placeholder: const Text('Select framework...'),
+          onSearchChanged: (value) => setState(() => searchValue = value),
+          searchPlaceholder: const Text('Search framework'),
+          options: [
+            if (filteredFrameworks.isEmpty)
+              const Padding(
+                padding: EdgeInsets.symmetric(vertical: 24),
+                child: Text('No framework found'),
+              ),
+            ...frameworks.entries.map(
+              (framework) {
+                // this offstage is used to avoid the focus loss when the search results appear again
+                // because it keeps the widget in the tree.
+                return Offstage(
+                  offstage: !filteredFrameworks.containsKey(framework.key),
+                  child: ShadOption(
+                    value: framework.key,
+                    child: Text(framework.value),
+                  ),
+                );
+              },
+            )
+          ],
+          selectedOptionBuilder: (context, value) => Text(frameworks[value]!),
         ),
       ],
     );
