@@ -3,7 +3,6 @@ import 'package:flutter_animate/flutter_animate.dart';
 import 'package:lucide_icons_flutter/lucide_icons.dart';
 import 'package:shadcn_ui/src/components/button.dart';
 import 'package:shadcn_ui/src/components/image.dart';
-import 'package:shadcn_ui/src/raw_components/same_width_column.dart';
 import 'package:shadcn_ui/src/theme/theme.dart';
 import 'package:shadcn_ui/src/theme/themes/shadows.dart';
 import 'package:shadcn_ui/src/utils/position.dart';
@@ -105,6 +104,10 @@ class ShadDialog extends StatelessWidget {
     this.titleTextAlign,
     this.descriptionTextAlign,
     this.alignment,
+    this.mainAxisAlignment,
+    this.crossAxisAlignment,
+    this.scrollable,
+    this.scrollPadding,
   }) : variant = ShadDialogVariant.primary;
 
   const ShadDialog.alert({
@@ -134,6 +137,10 @@ class ShadDialog extends StatelessWidget {
     this.titleTextAlign,
     this.descriptionTextAlign,
     this.alignment,
+    this.mainAxisAlignment,
+    this.crossAxisAlignment,
+    this.scrollable,
+    this.scrollPadding,
   }) : variant = ShadDialogVariant.alert;
 
   final Widget? title;
@@ -162,6 +169,10 @@ class ShadDialog extends StatelessWidget {
   final TextAlign? titleTextAlign;
   final TextAlign? descriptionTextAlign;
   final Alignment? alignment;
+  final MainAxisAlignment? mainAxisAlignment;
+  final CrossAxisAlignment? crossAxisAlignment;
+  final bool? scrollable;
+  final EdgeInsets? scrollPadding;
 
   @override
   Widget build(BuildContext context) {
@@ -235,110 +246,134 @@ class ShadDialog extends StatelessWidget {
     final effectiveAlignment =
         alignment ?? effectiveDialogTheme.alignment ?? Alignment.center;
 
+    final effectiveMainAxisAlignment = mainAxisAlignment ??
+        effectiveDialogTheme.mainAxisAlignment ??
+        MainAxisAlignment.start;
+
+    final effectiveCrossAxisAlignment = crossAxisAlignment ??
+        effectiveDialogTheme.crossAxisAlignment ??
+        CrossAxisAlignment.stretch;
+
+    final effectiveScrollable =
+        scrollable ?? effectiveDialogTheme.scrollable ?? true;
+
+    final effectiveScrollPadding = scrollPadding ??
+        effectiveDialogTheme.scrollPadding ??
+        MediaQuery.viewInsetsOf(context);
+
+    Widget dialog = ConstrainedBox(
+      constraints: effectiveConstraints,
+      child: ShadResponsiveBuilder(
+        builder: (context, breakpoint) {
+          final sm = breakpoint >= theme.breakpoints.small;
+
+          final effectiveActionsAxis = actionsAxis ??
+              effectiveDialogTheme.actionsAxis ??
+              (sm ? Axis.horizontal : Axis.vertical);
+
+          final effectiveActionsMainAxisSize = actionsMainAxisSize ??
+              effectiveDialogTheme.actionsMainAxisSize ??
+              MainAxisSize.max;
+
+          final effectiveActionsMainAxisAlignment = actionsMainAxisAlignment ??
+              effectiveDialogTheme.actionsMainAxisAlignment ??
+              MainAxisAlignment.end;
+
+          final effectiveActionsVerticalDirection = actionsVerticalDirection ??
+              effectiveDialogTheme.actionsVerticalDirection ??
+              (sm ? VerticalDirection.down : VerticalDirection.up);
+
+          final effectiveTitleTextAlign = titleTextAlign ??
+              effectiveDialogTheme.titleTextAlign ??
+              (sm ? TextAlign.start : TextAlign.center);
+
+          final effectiveDescriptionTextAlign = descriptionTextAlign ??
+              effectiveDialogTheme.descriptionTextAlign ??
+              (sm ? TextAlign.start : TextAlign.center);
+
+          Widget effectiveActions = Flex(
+            direction: effectiveActionsAxis,
+            mainAxisSize: effectiveActionsMainAxisSize,
+            mainAxisAlignment: effectiveActionsMainAxisAlignment,
+            verticalDirection: effectiveActionsVerticalDirection,
+            children: actions,
+          );
+
+          if (!sm && effectiveExpandActionsWhenTiny) {
+            effectiveActions = ShadTheme(
+              data: theme.copyWith(
+                primaryButtonTheme:
+                    theme.primaryButtonTheme.copyWith(width: double.infinity),
+                secondaryButtonTheme:
+                    theme.secondaryButtonTheme.copyWith(width: double.infinity),
+                outlineButtonTheme:
+                    theme.outlineButtonTheme.copyWith(width: double.infinity),
+                ghostButtonTheme:
+                    theme.ghostButtonTheme.copyWith(width: double.infinity),
+                destructiveButtonTheme: theme.destructiveButtonTheme
+                    .copyWith(width: double.infinity),
+              ),
+              child: effectiveActions,
+            );
+          }
+          return DecoratedBox(
+            decoration: BoxDecoration(
+              color: effectiveBackgroundColor,
+              borderRadius: (!sm && effectiveRemoveBorderRadiusWhenTiny)
+                  ? null
+                  : effectiveRadius,
+              border: effectiveBorder,
+              boxShadow: effectiveShadows,
+            ),
+            child: Stack(
+              children: [
+                Padding(
+                  padding: effectivePadding,
+                  child: Column(
+                    mainAxisSize: MainAxisSize.min,
+                    mainAxisAlignment: effectiveMainAxisAlignment,
+                    crossAxisAlignment: effectiveCrossAxisAlignment,
+                    children: [
+                      if (title != null)
+                        DefaultTextStyle(
+                          style: effectiveTitleStyle,
+                          textAlign: effectiveTitleTextAlign,
+                          child: title!,
+                        ),
+                      if (description != null)
+                        DefaultTextStyle(
+                          style: effectiveDescriptionStyle,
+                          textAlign: effectiveDescriptionTextAlign,
+                          child: description!,
+                        ),
+                      if (content != null)
+                        DefaultTextStyle(
+                          style: effectiveDescriptionStyle,
+                          child: content!,
+                        ),
+                      if (actions.isNotEmpty) effectiveActions,
+                    ].separatedBy(SizedBox(height: effectiveGap)),
+                  ),
+                ),
+                if (effectiveCloseIcon != null)
+                  effectiveCloseIcon.positionedWith(effectiveCloseIconPosition),
+              ],
+            ),
+          );
+        },
+      ),
+    );
+
+    if (effectiveScrollable) {
+      dialog = SingleChildScrollView(
+        padding: effectiveScrollPadding,
+        child: dialog,
+      );
+    }
+
     return Align(
       alignment: effectiveAlignment,
-      child: ConstrainedBox(
-        constraints: effectiveConstraints,
-        child: ShadResponsiveBuilder(
-          builder: (context, breakpoint) {
-            final sm = breakpoint >= theme.breakpoints.small;
-
-            final effectiveActionsAxis = actionsAxis ??
-                effectiveDialogTheme.actionsAxis ??
-                (sm ? Axis.horizontal : Axis.vertical);
-
-            final effectiveActionsMainAxisSize = actionsMainAxisSize ??
-                effectiveDialogTheme.actionsMainAxisSize ??
-                (sm ? MainAxisSize.min : MainAxisSize.max);
-
-            final effectiveActionsMainAxisAlignment =
-                actionsMainAxisAlignment ??
-                    effectiveDialogTheme.actionsMainAxisAlignment ??
-                    MainAxisAlignment.end;
-
-            final effectiveActionsVerticalDirection =
-                actionsVerticalDirection ??
-                    effectiveDialogTheme.actionsVerticalDirection ??
-                    (sm ? VerticalDirection.down : VerticalDirection.up);
-
-            final effectiveTitleTextAlign = titleTextAlign ??
-                effectiveDialogTheme.titleTextAlign ??
-                (sm ? TextAlign.start : TextAlign.center);
-
-            final effectiveDescriptionTextAlign = descriptionTextAlign ??
-                effectiveDialogTheme.descriptionTextAlign ??
-                (sm ? TextAlign.start : TextAlign.center);
-
-            Widget effectiveActions = Flex(
-              direction: effectiveActionsAxis,
-              mainAxisSize: effectiveActionsMainAxisSize,
-              mainAxisAlignment: effectiveActionsMainAxisAlignment,
-              verticalDirection: effectiveActionsVerticalDirection,
-              children: actions,
-            );
-
-            if (!sm && effectiveExpandActionsWhenTiny) {
-              effectiveActions = ShadTheme(
-                data: theme.copyWith(
-                  primaryButtonTheme:
-                      theme.primaryButtonTheme.copyWith(width: double.infinity),
-                  secondaryButtonTheme: theme.secondaryButtonTheme
-                      .copyWith(width: double.infinity),
-                  outlineButtonTheme:
-                      theme.outlineButtonTheme.copyWith(width: double.infinity),
-                  ghostButtonTheme:
-                      theme.ghostButtonTheme.copyWith(width: double.infinity),
-                  destructiveButtonTheme: theme.destructiveButtonTheme
-                      .copyWith(width: double.infinity),
-                ),
-                child: effectiveActions,
-              );
-            }
-            return DecoratedBox(
-              decoration: BoxDecoration(
-                color: effectiveBackgroundColor,
-                borderRadius: (!sm && effectiveRemoveBorderRadiusWhenTiny)
-                    ? null
-                    : effectiveRadius,
-                border: effectiveBorder,
-                boxShadow: effectiveShadows,
-              ),
-              child: Stack(
-                children: [
-                  Padding(
-                    padding: effectivePadding,
-                    child: ShadSameWidthColumn(
-                      children: [
-                        if (title != null)
-                          DefaultTextStyle(
-                            style: effectiveTitleStyle,
-                            textAlign: effectiveTitleTextAlign,
-                            child: title!,
-                          ),
-                        if (description != null)
-                          DefaultTextStyle(
-                            style: effectiveDescriptionStyle,
-                            textAlign: effectiveDescriptionTextAlign,
-                            child: description!,
-                          ),
-                        if (content != null)
-                          DefaultTextStyle(
-                            style: effectiveDescriptionStyle,
-                            child: content!,
-                          ),
-                        if (actions.isNotEmpty) effectiveActions,
-                      ].separatedBy(SizedBox(height: effectiveGap)),
-                    ),
-                  ),
-                  if (effectiveCloseIcon != null)
-                    effectiveCloseIcon
-                        .positionedWith(effectiveCloseIconPosition),
-                ],
-              ),
-            );
-          },
-        ),
-      ),
+      child: dialog,
     );
   }
 }
