@@ -18,16 +18,13 @@ class ShadCheckboxFormField extends ShadFormBuilderField<bool> {
     super.autovalidateMode,
     super.restorationId,
     required bool initialValue,
-    super.decoration,
     super.focusNode,
+    ShadDecoration? decoration,
     String? Function(bool)? validator,
-    ShadDecoration? inputDecoration,
-    BorderRadius? radius,
     double? size,
     Duration? duration,
     Widget? icon,
     Color? color,
-    double? borderWidth,
     Widget? inputLabel,
     Widget? inputSublabel,
     EdgeInsets? padding,
@@ -43,17 +40,15 @@ class ShadCheckboxFormField extends ShadFormBuilderField<bool> {
               onChanged: state.didChange,
               enabled: state.enabled,
               focusNode: state.focusNode,
-              decoration: inputDecoration,
-              radius: radius,
               size: size,
               duration: duration,
               icon: icon,
               color: color,
-              borderWidth: borderWidth,
               label: inputLabel,
               sublabel: inputSublabel,
               padding: padding,
               direction: direction,
+              decoration: decoration,
             );
           },
         );
