@@ -19,7 +19,6 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
     super.autovalidateMode,
     super.restorationId,
     super.initialValue,
-    super.decoration,
     super.focusNode,
     super.validator,
     required ShadSelectedOptionBuilder<T> selectedOptionBuilder,
@@ -28,12 +27,9 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
     bool closeOnTapOutside = true,
     double? minWidth,
     double? maxHeight,
-    ShadDecoration? inputDecoration,
+    ShadDecoration? decoration,
     Widget? trailing,
     EdgeInsets? padding,
-    Color? backgroundColor,
-    BorderRadius? radius,
-    Border? border,
     EdgeInsets? optionsPadding,
     bool? showScrollToTopChevron,
     bool? showScrollToBottomChevron,
@@ -54,12 +50,9 @@ class ShadSelectFormField<T> extends ShadFormBuilderField<T> {
               anchor: anchor,
               minWidth: minWidth,
               maxHeight: maxHeight,
-              decoration: inputDecoration,
+              decoration: decoration,
               trailing: trailing,
               padding: padding,
-              backgroundColor: backgroundColor,
-              radius: radius,
-              border: border,
               optionsPadding: optionsPadding,
               showScrollToTopChevron: showScrollToTopChevron,
               showScrollToBottomChevron: showScrollToBottomChevron,
