@@ -18,7 +18,7 @@ class ShadSwitchFormField extends ShadFormBuilderField<bool> {
     super.autovalidateMode,
     super.restorationId,
     required bool initialValue,
-    super.decoration,
+    ShadDecoration? decoration,
     super.focusNode,
     String? Function(bool)? validator,
     Color? thumbColor,
@@ -28,8 +28,6 @@ class ShadSwitchFormField extends ShadFormBuilderField<bool> {
     double? width,
     double? margin,
     Duration? duration,
-    BorderRadius? radius,
-    ShadDecoration? inputDecoration,
     Widget? inputLabel,
     Widget? inputSublabel,
     EdgeInsets? padding,
@@ -51,8 +49,7 @@ class ShadSwitchFormField extends ShadFormBuilderField<bool> {
               onChanged: state.didChange,
               enabled: state.enabled,
               focusNode: state.focusNode,
-              decoration: inputDecoration,
-              radius: radius,
+              decoration: decoration,
               duration: duration,
               label: inputLabel,
               sublabel: inputSublabel,
