@@ -2,15 +2,13 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart';
 import 'package:lucide_icons_flutter/lucide_icons.dart';
 import 'package:shadcn_ui/src/components/disabled.dart';
-import 'package:shadcn_ui/src/components/focused.dart';
-
 import 'package:shadcn_ui/src/components/image.dart';
+import 'package:shadcn_ui/src/components/input.dart';
 import 'package:shadcn_ui/src/components/popover.dart';
+import 'package:shadcn_ui/src/raw_components/focusable.dart';
 import 'package:shadcn_ui/src/raw_components/portal.dart';
-import 'package:shadcn_ui/src/raw_components/same_width_column.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
 import 'package:shadcn_ui/src/theme/components/select.dart';
 import 'package:shadcn_ui/src/theme/theme.dart';
@@ -21,6 +19,8 @@ typedef ShadSelectedOptionBuilder<T> = Widget Function(
   BuildContext context,
   T value,
 );
+
+enum ShadSelectType { primary, search }
 
 class ShadSelect<T> extends StatefulWidget {
   const ShadSelect({
@@ -38,15 +38,48 @@ class ShadSelect<T> extends StatefulWidget {
     this.decoration,
     this.trailing,
     this.padding,
-    this.backgroundColor,
-    this.radius,
-    this.border,
     this.optionsPadding,
     this.showScrollToBottomChevron,
     this.showScrollToTopChevron,
     this.scrollController,
     this.anchor,
-  });
+  })  : type = ShadSelectType.primary,
+        onSearchChanged = null,
+        searchDivider = null,
+        searchPlaceholder = null,
+        searchInputPrefix = null,
+        searchPadding = null,
+        search = null,
+        clearSearchOnClose = false;
+
+  const ShadSelect.withSearch({
+    super.key,
+    required this.options,
+    required this.selectedOptionBuilder,
+    required ValueChanged<String> this.onSearchChanged,
+    this.searchDivider,
+    this.searchInputPrefix,
+    this.searchPlaceholder,
+    this.searchPadding,
+    this.search,
+    this.clearSearchOnClose,
+    this.enabled = true,
+    this.placeholder,
+    this.initialValue,
+    this.onChanged,
+    this.focusNode,
+    this.closeOnTapOutside = true,
+    this.minWidth,
+    this.maxHeight,
+    this.decoration,
+    this.trailing,
+    this.padding,
+    this.optionsPadding,
+    this.showScrollToBottomChevron,
+    this.showScrollToTopChevron,
+    this.scrollController,
+    this.anchor,
+  }) : type = ShadSelectType.search;
 
   /// The callback that is called when the value of the [ShadSelect] changes.
   final ValueChanged<T>? onChanged;
@@ -96,17 +129,6 @@ class ShadSelect<T> extends StatefulWidget {
   /// `EdgeInsets.symmetric(horizontal: 12, vertical: 8)`.
   final EdgeInsets? padding;
 
-  /// The background color of the [ShadSelect], defaults to
-  /// the color scheme color.
-  final Color? backgroundColor;
-
-  /// The radius of the [ShadSelect], defaults to `ShadThemeData.radius`.
-  final BorderRadius? radius;
-
-  /// The border of the [ShadSelect], defaults to
-  /// `Border.all(color: kDefaultSelectBorderColor)`.
-  final Border? border;
-
   /// The padding of the options of the [ShadSelect], defaults to
   /// `EdgeInsets.all(4)`.
   final EdgeInsets? optionsPadding;
@@ -124,14 +146,50 @@ class ShadSelect<T> extends StatefulWidget {
   /// `ShadAnchorAutoPosition(verticalOffset: 24, preferBelow: true)`.
   final ShadAnchorBase? anchor;
 
-  static ShadSelectState<T> of<T>(BuildContext context) {
-    return maybeOf<T>(context)!;
+  /// The type of the [ShadSelect], defaults to `ShadSelectType.primary`.
+  final ShadSelectType type;
+
+  /// The callback that is called when the search value changes.
+  final ValueChanged<String>? onSearchChanged;
+
+  /// The widget that is displayed between the search input and the options.
+  final Widget? searchDivider;
+
+  /// The prefix of the search input.
+  final Widget? searchInputPrefix;
+
+  /// The placeholder of the search input.
+  final Widget? searchPlaceholder;
+
+  /// The padding of the search input.
+  final EdgeInsets? searchPadding;
+
+  /// A complete customizable search input.
+  final Widget? search;
+
+  /// Whether to clear the search input when the popover is closed, defaults to
+  /// `true`.
+  final bool? clearSearchOnClose;
+
+  static ShadSelectState<T> of<T>(BuildContext context, {bool listen = true}) {
+    return maybeOf<T>(context, listen: listen)!;
   }
 
-  static ShadSelectState<T>? maybeOf<T>(BuildContext context) {
-    return context
-        .dependOnInheritedWidgetOfExactType<ShadInheritedSelectContainer<T>>()
-        ?.data;
+  static ShadSelectState<T>? maybeOf<T>(
+    BuildContext context, {
+    bool listen = true,
+  }) {
+    if (listen) {
+      return context
+          .dependOnInheritedWidgetOfExactType<ShadInheritedSelectContainer<T>>()
+          ?.data;
+    }
+    final provider = context
+        .getElementForInheritedWidgetOfExactType<
+            ShadInheritedSelectContainer<T>>()
+        ?.widget;
+
+    return (provider as ShadInheritedSelectContainer<T>?)?.data;
   }
 
   @override
@@ -162,22 +220,6 @@ class ShadSelectState<T> extends State<ShadSelect<T>> {
     }
     if (widget.focusNode == null) internalFocusNode = FocusNode();
 
-    // Toggle the popover when the user presses enter
-    focusNode.onKeyEvent = (node, event) {
-      if (event.logicalKey == LogicalKeyboardKey.enter &&
-          event is KeyDownEvent) {
-        if (controller.isOpen) {
-          controller.hide();
-          focusNode.requestFocus();
-        } else {
-          FocusScope.of(context).unfocus();
-          controller.show();
-        }
-        return KeyEventResult.handled;
-      }
-      return KeyEventResult.ignored;
-    };
-
     // react to the scroll position
     scrollController.addListener(() {
       if (!scrollController.hasClients) return;
@@ -185,6 +227,21 @@ class ShadSelectState<T> extends State<ShadSelect<T>> {
           scrollController.offset < scrollController.position.maxScrollExtent;
       showScrollToTop.value = scrollController.offset > 0;
     });
+
+    if (widget.type == ShadSelectType.search) {
+      controller.addListener(() {
+        if (controller.isOpen) return;
+        final effectiveClearSearchOnClose = widget.clearSearchOnClose ??
+            ShadTheme.of(context, listen: false)
+                .selectTheme
+                .clearSearchOnClose ??
+            true;
+
+        if (effectiveClearSearchOnClose) {
+          widget.onSearchChanged?.call('');
+        }
+      });
+    }
   }
 
   @override
@@ -246,26 +303,23 @@ class ShadSelectState<T> extends State<ShadSelect<T>> {
   @override
   Widget build(BuildContext context) {
     assert(debugCheckHasShadTheme(context));
-
     final theme = ShadTheme.of(context);
-    final effectiveDecoration = widget.decoration ?? theme.decoration;
+
+    final effectiveDecoration =
+        (theme.selectTheme.decoration ?? const ShadDecoration())
+            .mergeWith(widget.decoration);
+
     final decorationHorizontalPadding =
         effectiveDecoration.border?.padding?.horizontal ?? 0.0;
+
     final effectivePadding = widget.padding ??
         theme.selectTheme.padding ??
         const EdgeInsets.symmetric(horizontal: 12, vertical: 8);
-    final effectiveBackgroundColor = widget.backgroundColor ??
-        theme.selectTheme.backgroundColor ??
-        theme.colorScheme.background;
-    final effectiveRadius =
-        widget.radius ?? theme.selectTheme.radius ?? theme.radius;
-    final effectiveBorder = widget.border ??
-        theme.selectTheme.border ??
-        Border.all(color: theme.colorScheme.input);
 
     final effectiveShowScrollToTopChevron = widget.showScrollToTopChevron ??
         theme.selectTheme.showScrollToTopChevron ??
         true;
+
     final effectiveShowScrollToBottomChevron =
         widget.showScrollToBottomChevron ??
             theme.selectTheme.showScrollToBottomChevron ??
@@ -273,9 +327,10 @@ class ShadSelectState<T> extends State<ShadSelect<T>> {
 
     final effectiveAnchor = widget.anchor ??
         theme.selectTheme.anchor ??
-        const ShadAnchorAuto(
-          verticalOffset: 24,
-          preferBelow: true,
+        const ShadAnchor(
+          overlayAlignment: Alignment.bottomLeft,
+          childAlignment: Alignment.topLeft,
+          offset: Offset.zero,
         );
 
     final Widget effectiveText;
@@ -308,152 +363,200 @@ class ShadSelectState<T> extends State<ShadSelect<T>> {
         theme.selectTheme.optionsPadding ??
         const EdgeInsets.all(4);
 
-    return LayoutBuilder(
-      builder: (context, constraints) {
-        final calculatedMinWidth =
-            max(effectiveMinWidth, constraints.minWidth) -
-                decorationHorizontalPadding;
-
-        final Widget select = ShadDisabled(
-          disabled: !widget.enabled,
-          child: ShadFocused(
-            canRequestFocus: widget.enabled,
-            focusNode: focusNode,
-            builder: (context, focused, child) {
-              return ShadDecorator(
-                focused: focused,
-                decoration: effectiveDecoration,
-                child: child!,
-              );
-            },
-            child: GestureDetector(
-              onTap: () {
-                FocusScope.of(context).unfocus();
-                controller.toggle();
-              },
-              child: Container(
-                padding: effectivePadding,
-                constraints: BoxConstraints(minWidth: calculatedMinWidth),
-                decoration: BoxDecoration(
-                  color: effectiveBackgroundColor,
-                  borderRadius: effectiveRadius,
-                  border: effectiveBorder,
+    final search = switch (widget.type) {
+      ShadSelectType.primary => null,
+      ShadSelectType.search => Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            widget.search ??
+                ShadInput(
+                  prefix: Padding(
+                    padding: const EdgeInsets.only(right: 8),
+                    child: Icon(
+                      LucideIcons.search,
+                      size: 16,
+                      color: theme.colorScheme.popoverForeground,
+                    ),
+                  ),
+                  padding: widget.searchPadding ??
+                      theme.selectTheme.searchPadding ??
+                      const EdgeInsets.all(12),
+                  placeholder: widget.searchPlaceholder,
+                  decoration: ShadDecoration.none,
+                  onChanged: widget.onSearchChanged,
                 ),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    Flexible(child: effectiveText),
-                    effectiveTrailing,
-                  ],
-                ),
-              ),
-            ),
-          ),
-        );
+            widget.searchDivider ?? const Divider(height: 1),
+          ],
+        ),
+    };
 
-        final scrollToTopChild = effectiveShowScrollToTopChevron
-            ? ValueListenableBuilder(
-                valueListenable: showScrollToTop,
-                builder: (context, show, child) {
-                  return AnimatedSwitcher(
-                    duration: const Duration(milliseconds: 300),
-                    child: show
-                        ? MouseRegion(
-                            onEnter: (_) {
-                              shouldAnimateToTop = true;
-                              animateToTop();
-                            },
-                            onExit: (_) => shouldAnimateToTop = false,
-                            child: Container(
-                              width: calculatedMinWidth,
-                              padding: const EdgeInsets.symmetric(vertical: 4),
-                              child: ShadImage.square(
-                                LucideIcons.chevronUp,
-                                size: 16,
-                                color: theme.colorScheme.popoverForeground,
-                              ),
-                            ),
-                          )
-                        : const SizedBox(),
+    return CallbackShortcuts(
+      bindings: {
+        const SingleActivator(LogicalKeyboardKey.enter): controller.toggle,
+        const SingleActivator(LogicalKeyboardKey.escape): controller.hide,
+      },
+      child: FocusTraversalGroup(
+        policy: WidgetOrderTraversalPolicy(),
+        child: LayoutBuilder(
+          builder: (context, constraints) {
+            final calculatedMinWidth =
+                max(effectiveMinWidth, constraints.minWidth) -
+                    decorationHorizontalPadding;
+            final Widget select = ShadDisabled(
+              disabled: !widget.enabled,
+              child: ShadFocusable(
+                canRequestFocus: widget.enabled,
+                focusNode: focusNode,
+                builder: (context, focused, child) {
+                  return ShadDecorator(
+                    focused: focused,
+                    decoration: effectiveDecoration,
+                    child: child,
                   );
                 },
-              )
-            : null;
-
-        final scrollToBottomChild = effectiveShowScrollToBottomChevron
-            ? ValueListenableBuilder(
-                valueListenable: showScrollToBottom,
-                builder: (context, show, child) {
-                  return AnimatedSwitcher(
-                    duration: const Duration(milliseconds: 300),
-                    child: show
-                        ? MouseRegion(
-                            onEnter: (_) {
-                              shouldAnimateToBottom = true;
-                              animateToBottom();
-                            },
-                            onExit: (_) => shouldAnimateToBottom = false,
-                            child: Container(
-                              width: calculatedMinWidth,
-                              padding: const EdgeInsets.symmetric(vertical: 4),
-                              child: ShadImage.square(
-                                LucideIcons.chevronDown,
-                                size: 16,
-                                color: theme.colorScheme.popoverForeground,
-                              ),
-                            ),
-                          )
-                        : const SizedBox(),
-                  );
-                },
-              )
-            : null;
-
-        return ShadInheritedSelectContainer(
-          data: this,
-          child: ShadPopover(
-            padding: EdgeInsets.zero,
-            controller: controller,
-            anchor: effectiveAnchor,
-            closeOnTapOutside: widget.closeOnTapOutside,
-            popover: (_) {
-              // set the initial value for showScrollToBottom and
-              // showScrollToTop, after the popover is rendered
-              WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
-                if (scrollController.hasClients) {
-                  showScrollToBottom.value = scrollController.offset <
-                      scrollController.position.maxScrollExtent;
-                  showScrollToTop.value = scrollController.offset > 0;
-                }
-              });
-
-              return ConstrainedBox(
-                constraints: BoxConstraints(maxHeight: effectiveMaxHeight),
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    if (scrollToTopChild != null) scrollToTopChild,
-                    Flexible(
-                      child: ConstrainedBox(
-                        constraints:
-                            BoxConstraints(minWidth: calculatedMinWidth),
-                        child: SingleChildScrollView(
-                          padding: effectiveOptionsPadding,
-                          controller: scrollController,
-                          child: ShadSameWidthColumn(children: widget.options),
-                        ),
+                child: ShadGestureDetector(
+                  cursor: SystemMouseCursors.click,
+                  behavior: HitTestBehavior.opaque,
+                  onTap: controller.toggle,
+                  child: ConstrainedBox(
+                    constraints: BoxConstraints(minWidth: calculatedMinWidth),
+                    child: Padding(
+                      padding: effectivePadding,
+                      child: Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                        mainAxisSize: MainAxisSize.min,
+                        children: [
+                          Flexible(child: effectiveText),
+                          effectiveTrailing,
+                        ],
                       ),
                     ),
-                    if (scrollToBottomChild != null) scrollToBottomChild,
-                  ],
+                  ),
                 ),
-              );
-            },
-            child: select,
-          ),
-        );
-      },
+              ),
+            );
+
+            final scrollToTopChild = effectiveShowScrollToTopChevron
+                ? ValueListenableBuilder(
+                    valueListenable: showScrollToTop,
+                    builder: (context, show, child) {
+                      return AnimatedSwitcher(
+                        duration: const Duration(milliseconds: 300),
+                        child: show
+                            ? MouseRegion(
+                                onEnter: (_) {
+                                  shouldAnimateToTop = true;
+                                  animateToTop();
+                                },
+                                onExit: (_) => shouldAnimateToTop = false,
+                                child: Container(
+                                  width: calculatedMinWidth,
+                                  padding:
+                                      const EdgeInsets.symmetric(vertical: 4),
+                                  child: ShadImage.square(
+                                    LucideIcons.chevronUp,
+                                    size: 16,
+                                    color: theme.colorScheme.popoverForeground,
+                                  ),
+                                ),
+                              )
+                            : const SizedBox(),
+                      );
+                    },
+                  )
+                : null;
+
+            final scrollToBottomChild = effectiveShowScrollToBottomChevron
+                ? ValueListenableBuilder(
+                    valueListenable: showScrollToBottom,
+                    builder: (context, show, child) {
+                      return AnimatedSwitcher(
+                        duration: const Duration(milliseconds: 300),
+                        child: show
+                            ? MouseRegion(
+                                onEnter: (_) {
+                                  shouldAnimateToBottom = true;
+                                  animateToBottom();
+                                },
+                                onExit: (_) => shouldAnimateToBottom = false,
+                                child: Container(
+                                  width: calculatedMinWidth,
+                                  padding:
+                                      const EdgeInsets.symmetric(vertical: 4),
+                                  child: ShadImage.square(
+                                    LucideIcons.chevronDown,
+                                    size: 16,
+                                    color: theme.colorScheme.popoverForeground,
+                                  ),
+                                ),
+                              )
+                            : const SizedBox(),
+                      );
+                    },
+                  )
+                : null;
+
+            return ShadInheritedSelectContainer(
+              data: this,
+              child: ShadPopover(
+                padding: EdgeInsets.zero,
+                controller: controller,
+                anchor: effectiveAnchor,
+                closeOnTapOutside: widget.closeOnTapOutside,
+                popover: (_) {
+                  // set the initial value for showScrollToBottom and
+                  // showScrollToTop, after the popover is rendered
+                  WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
+                    if (scrollController.hasClients) {
+                      showScrollToBottom.value = scrollController.offset <
+                          scrollController.position.maxScrollExtent;
+                      showScrollToTop.value = scrollController.offset > 0;
+                    }
+                  });
+
+                  return ConstrainedBox(
+                    constraints: BoxConstraints(maxHeight: effectiveMaxHeight),
+                    child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      children: [
+                        if (search != null)
+                          Flexible(
+                            child: ConstrainedBox(
+                              constraints:
+                                  BoxConstraints(minWidth: calculatedMinWidth),
+                              child: IntrinsicWidth(child: search),
+                            ),
+                          ),
+                        if (scrollToTopChild != null) scrollToTopChild,
+                        Flexible(
+                          child: ConstrainedBox(
+                            constraints:
+                                BoxConstraints(minWidth: calculatedMinWidth),
+                            child: SingleChildScrollView(
+                              padding: effectiveOptionsPadding,
+                              controller: scrollController,
+                              child: IntrinsicWidth(
+                                child: Column(
+                                  crossAxisAlignment:
+                                      CrossAxisAlignment.stretch,
+                                  mainAxisSize: MainAxisSize.min,
+                                  children: widget.options.toList(),
+                                ),
+                              ),
+                            ),
+                          ),
+                        ),
+                        if (scrollToBottomChild != null) scrollToBottomChild,
+                      ],
+                    ),
+                  );
+                },
+                child: select,
+              ),
+            );
+          },
+        ),
+      ),
     );
   }
 }
@@ -511,10 +614,26 @@ class _ShadOptionState<T> extends State<ShadOption<T>> {
   final focusNode = FocusNode();
 
   @override
+  void initState() {
+    super.initState();
+    focusNode.addListener(onFocusChange);
+
+    final inherited = ShadSelect.of<T>(context, listen: false);
+    final selected = inherited.selected == widget.value;
+    if (selected) focusNode.requestFocus();
+  }
+
+  @override
   void dispose() {
-    focusNode.dispose();
+    focusNode
+      ..removeListener(onFocusChange)
+      ..dispose();
     hovered.dispose();
     super.dispose();
+  }
+
+  void onFocusChange() {
+    hovered.value = focusNode.hasFocus;
   }
 
   @override
@@ -526,6 +645,7 @@ class _ShadOptionState<T> extends State<ShadOption<T>> {
     );
     final inheritedSelect = ShadSelect.of<T>(context);
     final selected = inheritedSelect.selected == widget.value;
+
     if (selected) {
       // scroll to the selected option
       WidgetsBinding.instance.addPostFrameCallback((_) {
@@ -545,46 +665,55 @@ class _ShadOptionState<T> extends State<ShadOption<T>> {
     final effectiveRadius =
         widget.radius ?? theme.optionTheme.radius ?? theme.radius;
 
-    return Focus(
-      focusNode: focusNode,
-      canRequestFocus: true,
-      child: ShadGestureDetector(
-        onHoverChange: (value) => hovered.value = value,
-        onTap: () => inheritedSelect.select(widget.value),
-        child: ValueListenableBuilder(
-          valueListenable: hovered,
-          builder: (context, hovered, child) {
-            return Container(
-              padding: effectivePadding,
-              decoration: BoxDecoration(
-                color: hovered ? effectiveHoveredBackgroundColor : null,
-                borderRadius: effectiveRadius,
-              ),
-              child: child,
-            );
+    return CallbackShortcuts(
+      bindings: {
+        const SingleActivator(LogicalKeyboardKey.enter): () {
+          inheritedSelect.select(widget.value);
+        },
+      },
+      child: Focus(
+        focusNode: focusNode,
+        child: ShadGestureDetector(
+          behavior: HitTestBehavior.opaque,
+          cursor: SystemMouseCursors.click,
+          onHoverChange: (value) {
+            hovered.value = value;
           },
-          child: Row(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              widget.selectedIcon ??
-                  Visibility.maintain(
-                    visible: selected,
-                    child: Padding(
-                      padding: const EdgeInsets.only(right: 8),
-                      child: ShadImage.square(
-                        LucideIcons.check,
-                        size: 16,
-                        color: theme.colorScheme.popoverForeground,
+          onTap: () => inheritedSelect.select(widget.value),
+          child: ValueListenableBuilder(
+            valueListenable: hovered,
+            builder: (context, hovered, child) {
+              return Container(
+                padding: effectivePadding,
+                decoration: BoxDecoration(
+                  color: hovered ? effectiveHoveredBackgroundColor : null,
+                  borderRadius: effectiveRadius,
+                ),
+                child: child,
+              );
+            },
+            child: Row(
+              children: [
+                widget.selectedIcon ??
+                    Visibility.maintain(
+                      visible: selected,
+                      child: Padding(
+                        padding: const EdgeInsets.only(right: 8),
+                        child: ShadImage.square(
+                          LucideIcons.check,
+                          size: 16,
+                          color: theme.colorScheme.popoverForeground,
+                        ),
                       ),
                     ),
+                DefaultTextStyle(
+                  style: theme.textTheme.muted.copyWith(
+                    color: theme.colorScheme.popoverForeground,
                   ),
-              DefaultTextStyle(
-                style: theme.textTheme.muted.copyWith(
-                  color: theme.colorScheme.popoverForeground,
+                  child: widget.child,
                 ),
-                child: widget.child,
-              ),
-            ],
+              ],
+            ),
           ),
         ),
       ),
