@@ -9,24 +9,25 @@ class ShadAccordionTheme {
     this.merge = true,
     this.iconSrc,
     this.iconEffects,
-    this.transitionBuilder,
     this.padding,
     this.underlineTitleOnHover,
     this.titleStyle,
     this.curve,
     this.duration,
+    this.maintainState,
+    this.effects,
   });
 
   final bool merge;
   final ShadImageSrc? iconSrc;
   final List<Effect<dynamic>>? iconEffects;
-  final Widget Function(Animation<double> animation, Widget child)?
-      transitionBuilder;
   final EdgeInsets? padding;
   final bool? underlineTitleOnHover;
   final TextStyle? titleStyle;
   final Curve? curve;
   final Duration? duration;
+  final bool? maintainState;
+  final List<Effect<dynamic>>? effects;
 
   static ShadAccordionTheme lerp(
     ShadAccordionTheme a,
@@ -35,15 +36,17 @@ class ShadAccordionTheme {
   ) {
     if (identical(a, b)) return a;
     return ShadAccordionTheme(
-      merge: b.merge,
-      iconSrc: b.iconSrc,
-      iconEffects: b.iconEffects,
-      transitionBuilder: b.transitionBuilder,
+      merge: t < 0.5 ? a.merge : b.merge,
+      iconSrc: t < 0.5 ? a.iconSrc : b.iconSrc,
+      iconEffects: t < 0.5 ? a.iconEffects : b.iconEffects,
       padding: EdgeInsets.lerp(a.padding, b.padding, t),
-      underlineTitleOnHover: b.underlineTitleOnHover,
+      underlineTitleOnHover:
+          t < 0.5 ? a.underlineTitleOnHover : b.underlineTitleOnHover,
       titleStyle: TextStyle.lerp(a.titleStyle, b.titleStyle, t),
-      curve: b.curve,
-      duration: b.duration,
+      curve: t < 0.5 ? a.curve : b.curve,
+      duration: t < 0.5 ? a.duration : b.duration,
+      maintainState: t < 0.5 ? a.maintainState : b.maintainState,
+      effects: t < 0.5 ? a.effects : b.effects,
     );
   }
 
@@ -58,6 +61,8 @@ class ShadAccordionTheme {
       titleStyle: other.titleStyle,
       curve: other.curve,
       duration: other.duration,
+      maintainState: other.maintainState,
+      effects: other.effects,
     );
   }
 
@@ -73,7 +78,9 @@ class ShadAccordionTheme {
         other.underlineTitleOnHover == underlineTitleOnHover &&
         other.titleStyle == titleStyle &&
         other.curve == curve &&
-        other.duration == duration;
+        other.duration == duration &&
+        other.maintainState == maintainState &&
+        listEquals(other.effects, effects);
   }
 
   @override
@@ -85,7 +92,9 @@ class ShadAccordionTheme {
         underlineTitleOnHover.hashCode ^
         titleStyle.hashCode ^
         curve.hashCode ^
-        duration.hashCode;
+        duration.hashCode ^
+        maintainState.hashCode ^
+        effects.hashCode;
   }
 
   ShadAccordionTheme copyWith({
@@ -97,6 +106,8 @@ class ShadAccordionTheme {
     TextStyle? titleStyle,
     Curve? curve,
     Duration? duration,
+    bool? maintainState,
+    List<Effect<dynamic>>? effects,
   }) {
     return ShadAccordionTheme(
       merge: merge ?? this.merge,
@@ -108,6 +119,8 @@ class ShadAccordionTheme {
       titleStyle: titleStyle ?? this.titleStyle,
       curve: curve ?? this.curve,
       duration: duration ?? this.duration,
+      maintainState: maintainState ?? this.maintainState,
+      effects: effects ?? this.effects,
     );
   }
 }
