@@ -7,15 +7,25 @@ class ShadInputTheme {
     this.merge = true,
     this.decoration,
     this.padding,
-    this.border,
-    this.radius,
+    this.backgroundColor,
+    this.style,
+    this.placeholderStyle,
+    this.placeholderAlignment,
+    this.inputPadding,
+    this.mainAxisAlignment,
+    this.crossAxisAlignment,
   });
 
   final bool merge;
   final ShadDecoration? decoration;
   final EdgeInsets? padding;
-  final Border? border;
-  final BorderRadius? radius;
+  final Color? backgroundColor;
+  final TextStyle? style;
+  final TextStyle? placeholderStyle;
+  final Alignment? placeholderAlignment;
+  final EdgeInsets? inputPadding;
+  final MainAxisAlignment? mainAxisAlignment;
+  final CrossAxisAlignment? crossAxisAlignment;
 
   static ShadInputTheme lerp(
     ShadInputTheme a,
@@ -27,8 +37,13 @@ class ShadInputTheme {
       merge: b.merge,
       decoration: ShadDecoration.lerp(a.decoration, b.decoration, t),
       padding: EdgeInsets.lerp(a.padding, b.padding, t),
-      border: Border.lerp(a.border, b.border, t),
-      radius: BorderRadius.lerp(a.radius, b.radius, t),
+      backgroundColor: Color.lerp(a.backgroundColor, b.backgroundColor, t),
+      style: TextStyle.lerp(a.style, b.style, t),
+      placeholderStyle:
+          TextStyle.lerp(a.placeholderStyle, b.placeholderStyle, t),
+      inputPadding: EdgeInsets.lerp(a.inputPadding, b.inputPadding, t),
+      mainAxisAlignment: t < 0.5 ? a.mainAxisAlignment : b.mainAxisAlignment,
+      crossAxisAlignment: t < 0.5 ? a.crossAxisAlignment : b.crossAxisAlignment,
     );
   }
 
@@ -36,15 +51,23 @@ class ShadInputTheme {
     bool? merge,
     ShadDecoration? decoration,
     EdgeInsets? padding,
-    Border? border,
-    BorderRadius? radius,
+    Color? backgroundColor,
+    TextStyle? style,
+    TextStyle? placeholderStyle,
+    EdgeInsets? inputPadding,
+    MainAxisAlignment? mainAxisAlignment,
+    CrossAxisAlignment? crossAxisAlignment,
   }) {
     return ShadInputTheme(
       merge: merge ?? this.merge,
       decoration: decoration ?? this.decoration,
       padding: padding ?? this.padding,
-      border: border ?? this.border,
-      radius: radius ?? this.radius,
+      backgroundColor: backgroundColor ?? this.backgroundColor,
+      style: style ?? this.style,
+      placeholderStyle: placeholderStyle ?? this.placeholderStyle,
+      inputPadding: inputPadding ?? this.inputPadding,
+      mainAxisAlignment: mainAxisAlignment ?? this.mainAxisAlignment,
+      crossAxisAlignment: crossAxisAlignment ?? this.crossAxisAlignment,
     );
   }
 
@@ -52,10 +75,14 @@ class ShadInputTheme {
     if (other == null) return this;
     if (!other.merge) return other;
     return copyWith(
-      decoration: other.decoration,
+      decoration: decoration?.mergeWith(other.decoration) ?? other.decoration,
       padding: other.padding,
-      border: other.border,
-      radius: other.radius,
+      backgroundColor: other.backgroundColor,
+      style: other.style,
+      placeholderStyle: other.placeholderStyle,
+      inputPadding: other.inputPadding,
+      mainAxisAlignment: other.mainAxisAlignment,
+      crossAxisAlignment: other.crossAxisAlignment,
     );
   }
 
@@ -67,8 +94,12 @@ class ShadInputTheme {
         other.merge == merge &&
         other.decoration == decoration &&
         other.padding == padding &&
-        other.border == border &&
-        other.radius == radius;
+        other.backgroundColor == backgroundColor &&
+        other.style == style &&
+        other.placeholderStyle == placeholderStyle &&
+        other.inputPadding == inputPadding &&
+        other.mainAxisAlignment == mainAxisAlignment &&
+        other.crossAxisAlignment == crossAxisAlignment;
   }
 
   @override
@@ -76,7 +107,11 @@ class ShadInputTheme {
     return merge.hashCode ^
         decoration.hashCode ^
         padding.hashCode ^
-        border.hashCode ^
-        radius.hashCode;
+        backgroundColor.hashCode ^
+        style.hashCode ^
+        placeholderStyle.hashCode ^
+        inputPadding.hashCode ^
+        mainAxisAlignment.hashCode ^
+        crossAxisAlignment.hashCode;
   }
 }
