@@ -16,12 +16,11 @@ class ShadSelectTheme {
     this.maxHeight,
     this.padding,
     this.optionsPadding,
-    this.backgroundColor,
-    this.radius,
-    this.border,
     this.showScrollToTopChevron,
     this.showScrollToBottomChevron,
     this.anchor,
+    this.searchPadding,
+    this.clearSearchOnClose,
   });
 
   final bool merge;
@@ -30,12 +29,11 @@ class ShadSelectTheme {
   final double? maxHeight;
   final EdgeInsets? padding;
   final EdgeInsets? optionsPadding;
-  final Color? backgroundColor;
-  final BorderRadius? radius;
-  final Border? border;
   final bool? showScrollToTopChevron;
   final bool? showScrollToBottomChevron;
   final ShadAnchorBase? anchor;
+  final EdgeInsets? searchPadding;
+  final bool? clearSearchOnClose;
 
   static ShadSelectTheme lerp(
     ShadSelectTheme a,
@@ -49,13 +47,12 @@ class ShadSelectTheme {
       minWidth: lerpDouble(a.minWidth, b.minWidth, t),
       maxHeight: lerpDouble(a.maxHeight, b.maxHeight, t),
       padding: EdgeInsets.lerp(a.padding, b.padding, t),
-      backgroundColor: Color.lerp(a.backgroundColor, b.backgroundColor, t),
-      radius: BorderRadius.lerp(a.radius, b.radius, t),
-      border: Border.lerp(a.border, b.border, t),
       optionsPadding: EdgeInsets.lerp(a.optionsPadding, b.optionsPadding, t),
       showScrollToTopChevron: b.showScrollToTopChevron,
       showScrollToBottomChevron: b.showScrollToBottomChevron,
       anchor: b.anchor,
+      searchPadding: EdgeInsets.lerp(a.searchPadding, b.searchPadding, t),
+      clearSearchOnClose: t < 0.5 ? a.clearSearchOnClose : b.clearSearchOnClose,
     );
   }
 
@@ -65,13 +62,12 @@ class ShadSelectTheme {
     double? minWidth,
     double? maxHeight,
     EdgeInsets? padding,
-    Color? backgroundColor,
-    BorderRadius? radius,
-    Border? border,
     EdgeInsets? optionsPadding,
     bool? showScrollToTopChevron,
     bool? showScrollToBottomChevron,
     ShadAnchorBase? anchor,
+    EdgeInsets? searchPadding,
+    bool? clearSearchOnClose,
   }) {
     return ShadSelectTheme(
       merge: merge ?? this.merge,
@@ -79,15 +75,14 @@ class ShadSelectTheme {
       minWidth: minWidth ?? this.minWidth,
       maxHeight: maxHeight ?? this.maxHeight,
       padding: padding ?? this.padding,
-      backgroundColor: backgroundColor ?? this.backgroundColor,
-      radius: radius ?? this.radius,
-      border: border ?? this.border,
       optionsPadding: optionsPadding ?? this.optionsPadding,
       showScrollToTopChevron:
           showScrollToTopChevron ?? this.showScrollToTopChevron,
       showScrollToBottomChevron:
           showScrollToBottomChevron ?? this.showScrollToBottomChevron,
       anchor: anchor ?? this.anchor,
+      searchPadding: searchPadding ?? this.searchPadding,
+      clearSearchOnClose: clearSearchOnClose ?? this.clearSearchOnClose,
     );
   }
 
@@ -95,17 +90,16 @@ class ShadSelectTheme {
     if (other == null) return this;
     if (!other.merge) return other;
     return copyWith(
-      decoration: other.decoration,
+      decoration: decoration?.mergeWith(other.decoration) ?? other.decoration,
       minWidth: other.minWidth,
       maxHeight: other.maxHeight,
       padding: other.padding,
-      backgroundColor: other.backgroundColor,
-      radius: other.radius,
-      border: other.border,
       optionsPadding: other.optionsPadding,
       showScrollToTopChevron: other.showScrollToTopChevron,
       showScrollToBottomChevron: other.showScrollToBottomChevron,
       anchor: other.anchor,
+      searchPadding: other.searchPadding,
+      clearSearchOnClose: other.clearSearchOnClose,
     );
   }
 
@@ -119,13 +113,12 @@ class ShadSelectTheme {
         other.minWidth == minWidth &&
         other.maxHeight == maxHeight &&
         other.padding == padding &&
-        other.backgroundColor == backgroundColor &&
-        other.radius == radius &&
-        other.border == border &&
         other.optionsPadding == optionsPadding &&
         other.showScrollToTopChevron == showScrollToTopChevron &&
         other.showScrollToBottomChevron == showScrollToBottomChevron &&
-        other.anchor == anchor;
+        other.anchor == anchor &&
+        other.searchPadding == searchPadding &&
+        other.clearSearchOnClose == clearSearchOnClose;
   }
 
   @override
@@ -135,12 +128,11 @@ class ShadSelectTheme {
         minWidth.hashCode ^
         maxHeight.hashCode ^
         padding.hashCode ^
-        backgroundColor.hashCode ^
-        radius.hashCode ^
-        border.hashCode ^
         optionsPadding.hashCode ^
         showScrollToTopChevron.hashCode ^
         showScrollToBottomChevron.hashCode ^
-        anchor.hashCode;
+        anchor.hashCode ^
+        searchPadding.hashCode ^
+        clearSearchOnClose.hashCode;
   }
 }
