@@ -1,3 +1,4 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_animate/flutter_animate.dart';
@@ -29,62 +30,72 @@ import 'package:shadcn_ui/src/theme/components/tooltip.dart';
 import 'package:shadcn_ui/src/theme/text_theme/text_styles_default.dart';
 import 'package:shadcn_ui/src/theme/text_theme/theme.dart';
 import 'package:shadcn_ui/src/theme/themes/shadows.dart';
-import 'package:shadcn_ui/src/utils/extensions.dart';
 import 'package:shadcn_ui/src/utils/gesture_detector.dart';
 import 'package:shadcn_ui/src/utils/position.dart';
 
-abstract class ShadComponentDefaultTheme {
+abstract class ShadDefaultComponentThemes {
   static ShadButtonTheme primaryButtonTheme({
     required ShadColorScheme colorScheme,
+    required BorderRadius radius,
   }) {
     return ShadButtonTheme(
       backgroundColor: colorScheme.primary,
       hoverBackgroundColor: colorScheme.primary.withOpacity(.9),
       foregroundColor: colorScheme.primaryForeground,
       hoverForegroundColor: colorScheme.primaryForeground,
+      decoration: ShadDecoration(border: ShadBorder(radius: radius)),
     );
   }
 
   static ShadButtonTheme secondaryButtonTheme({
     required ShadColorScheme colorScheme,
+    required BorderRadius radius,
   }) {
     return ShadButtonTheme(
       backgroundColor: colorScheme.secondary,
       hoverBackgroundColor: colorScheme.secondary.withOpacity(.8),
       foregroundColor: colorScheme.secondaryForeground,
       hoverForegroundColor: colorScheme.secondaryForeground,
+      decoration: ShadDecoration(border: ShadBorder(radius: radius)),
     );
   }
 
   static ShadButtonTheme destructiveButtonTheme({
     required ShadColorScheme colorScheme,
+    required BorderRadius radius,
   }) {
     return ShadButtonTheme(
       backgroundColor: colorScheme.destructive,
       hoverBackgroundColor: colorScheme.destructive.withOpacity(.9),
       foregroundColor: colorScheme.destructiveForeground,
       hoverForegroundColor: colorScheme.destructiveForeground,
+      decoration: ShadDecoration(border: ShadBorder(radius: radius)),
     );
   }
 
   static ShadButtonTheme outlineButtonTheme({
     required ShadColorScheme colorScheme,
+    required BorderRadius radius,
   }) {
     return ShadButtonTheme(
       hoverBackgroundColor: colorScheme.accent,
       foregroundColor: colorScheme.primary,
       hoverForegroundColor: colorScheme.accentForeground,
-      border: Border.all(color: colorScheme.input),
+      decoration: ShadDecoration(
+        border: ShadBorder(radius: radius, color: colorScheme.input),
+      ),
     );
   }
 
   static ShadButtonTheme ghostButtonTheme({
     required ShadColorScheme colorScheme,
+    required BorderRadius radius,
   }) {
     return ShadButtonTheme(
       hoverBackgroundColor: colorScheme.accent,
       foregroundColor: colorScheme.primary,
       hoverForegroundColor: colorScheme.accentForeground,
+      decoration: ShadDecoration(border: ShadBorder(radius: radius)),
     );
   }
 
@@ -186,15 +197,11 @@ abstract class ShadComponentDefaultTheme {
         ScaleEffect(begin: Offset(.95, .95), end: Offset(1, 1)),
         MoveEffect(begin: Offset(0, 2), end: Offset.zero),
       ],
-      shadows: ShadShadows.md,
       padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 6),
-      decoration: BoxDecoration(
-        borderRadius: radius,
-        border: Border.all(
-          color: colorScheme.border,
-        ),
+      decoration: ShadDecoration(
+        border: ShadBorder(radius: radius, color: colorScheme.border),
         color: colorScheme.popover,
-        boxShadow: ShadShadows.md,
+        shadows: ShadShadows.md,
       ),
       anchor: const ShadAnchorAuto(
         preferBelow: false,
@@ -215,25 +222,31 @@ abstract class ShadComponentDefaultTheme {
       ],
       shadows: ShadShadows.md,
       padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 6),
-      decoration: BoxDecoration(
-        borderRadius: radius,
-        border: Border.all(
+      decoration: ShadDecoration(
+        color: colorScheme.popover,
+        shadows: ShadShadows.md,
+        border: ShadBorder(
+          radius: radius,
           color: colorScheme.border,
         ),
-        color: colorScheme.popover,
-        boxShadow: ShadShadows.md,
       ),
       anchor: const ShadAnchorAuto(preferBelow: true, verticalOffset: 24),
     );
   }
 
-  static ShadDecoration decoration({
+  static ShadDecoration decorationTheme({
     required ShadColorScheme colorScheme,
     required BorderRadius radius,
     required ShadTextTheme textTheme,
   }) {
     return ShadDecoration(
-      border: const ShadBorder(padding: EdgeInsets.all(4)),
+      secondaryBorder: const ShadBorder(padding: EdgeInsets.all(4)),
+      secondaryFocusedBorder: ShadBorder(
+        width: 2,
+        color: colorScheme.ring,
+        radius: radius.add(radius / 2),
+        padding: const EdgeInsets.all(2),
+      ),
       labelStyle: textTheme.muted.copyWith(
         fontWeight: FontWeight.w500,
         color: colorScheme.foreground,
@@ -250,89 +263,49 @@ abstract class ShadComponentDefaultTheme {
         fontWeight: FontWeight.w500,
         color: colorScheme.destructive,
       ),
-      focusedBorder: ShadBorder(
-        width: 2,
-        color: colorScheme.ring,
-        radius: radius.add(radius / 2),
-        padding: const EdgeInsets.all(2),
-      ),
     );
   }
 
-  static ShadTextTheme textTheme({
-    required ShadColorScheme colorScheme,
-  }) {
+  static ShadTextTheme textTheme() {
     const defaultFontFamily = 'packages/shadcn_ui/$kDefaultFontFamily';
     return ShadTextTheme.custom(
-      h1Large: ShadTextDefaultTheme.h1Large(
-        colorScheme: colorScheme,
-        family: defaultFontFamily,
-      ),
-      h1: ShadTextDefaultTheme.h1(
-        colorScheme: colorScheme,
-        family: defaultFontFamily,
-      ),
-      h2: ShadTextDefaultTheme.h2(
-        colorScheme: colorScheme,
-        family: defaultFontFamily,
-      ),
-      h3: ShadTextDefaultTheme.h3(
-        colorScheme: colorScheme,
-        family: defaultFontFamily,
-      ),
-      h4: ShadTextDefaultTheme.h4(
-        colorScheme: colorScheme,
-        family: defaultFontFamily,
-      ),
-      p: ShadTextDefaultTheme.p(
-        colorScheme: colorScheme,
-        family: defaultFontFamily,
-      ),
-      blockquote: ShadTextDefaultTheme.blockquote(
-        colorScheme: colorScheme,
-        family: defaultFontFamily,
-      ),
-      table: ShadTextDefaultTheme.table(
-        colorScheme: colorScheme,
-        family: defaultFontFamily,
-      ),
-      list: ShadTextDefaultTheme.list(
-        colorScheme: colorScheme,
-        family: defaultFontFamily,
-      ),
-      lead: ShadTextDefaultTheme.lead(
-        colorScheme: colorScheme,
-        family: defaultFontFamily,
-      ),
-      large: ShadTextDefaultTheme.large(
-        colorScheme: colorScheme,
-        family: defaultFontFamily,
-      ),
-      small: ShadTextDefaultTheme.small(
-        colorScheme: colorScheme,
-        family: defaultFontFamily,
-      ),
-      muted: ShadTextDefaultTheme.muted(
-        colorScheme: colorScheme,
-        family: defaultFontFamily,
-      ),
+      h1Large: ShadTextDefaultTheme.h1Large(family: defaultFontFamily),
+      h1: ShadTextDefaultTheme.h1(family: defaultFontFamily),
+      h2: ShadTextDefaultTheme.h2(family: defaultFontFamily),
+      h3: ShadTextDefaultTheme.h3(family: defaultFontFamily),
+      h4: ShadTextDefaultTheme.h4(family: defaultFontFamily),
+      p: ShadTextDefaultTheme.p(family: defaultFontFamily),
+      blockquote: ShadTextDefaultTheme.blockquote(family: defaultFontFamily),
+      table: ShadTextDefaultTheme.table(family: defaultFontFamily),
+      list: ShadTextDefaultTheme.list(family: defaultFontFamily),
+      lead: ShadTextDefaultTheme.lead(family: defaultFontFamily),
+      large: ShadTextDefaultTheme.large(family: defaultFontFamily),
+      small: ShadTextDefaultTheme.small(family: defaultFontFamily),
+      muted: ShadTextDefaultTheme.muted(family: defaultFontFamily),
       family: defaultFontFamily,
     );
   }
 
   static ShadSelectTheme selectTheme({
     required ShadColorScheme colorScheme,
+    required BorderRadius radius,
   }) {
     return ShadSelectTheme(
       minWidth: kDefaultSelectMinWidth,
       maxHeight: kDefaultSelectMaxHeight,
       padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-      backgroundColor: colorScheme.background,
-      border: Border.all(color: colorScheme.input),
+      decoration: ShadDecoration(
+        border: ShadBorder(radius: radius, color: colorScheme.input),
+      ),
       optionsPadding: const EdgeInsets.all(4),
       showScrollToTopChevron: true,
       showScrollToBottomChevron: true,
-      anchor: const ShadAnchorAuto(preferBelow: true, verticalOffset: 24),
+      anchor: const ShadAnchor(
+        overlayAlignment: Alignment.bottomLeft,
+        childAlignment: Alignment.topLeft,
+        offset: Offset.zero,
+      ),
+      searchPadding: const EdgeInsets.all(12),
     );
   }
 
@@ -366,49 +339,39 @@ abstract class ShadComponentDefaultTheme {
   static ShadSwitchTheme switchTheme({
     required ShadColorScheme colorScheme,
   }) {
+    const width = 44.0;
+    const height = 24.0;
+    const radius = BorderRadius.all(Radius.circular(height / 2));
     return ShadSwitchTheme(
-      radius: const BorderRadius.all(Radius.circular(24)),
-      width: 44,
+      width: width,
+      height: height,
       margin: 2,
       duration: 100.milliseconds,
       thumbColor: colorScheme.background,
       uncheckedTrackColor: colorScheme.input,
       checkedTrackColor: colorScheme.primary,
+      padding: const EdgeInsets.only(left: 8),
       decoration: ShadDecoration(
-        border: const ShadBorder(width: 4, color: Colors.transparent),
-        focusedBorder: ShadBorder(
-          width: 2,
-          color: colorScheme.ring,
-          // the outer radius is calculated as
-          // outerRadius = innerRadius + innerRadius / 2
-          // outerRadius = 24 + 24 / 2 = 36
-          radius: const BorderRadius.all(Radius.circular(36)),
-          padding: const EdgeInsets.all(2),
-        ),
+        border: ShadBorder(radius: radius.add(radius / 2)),
+        secondaryFocusedBorder: ShadBorder(radius: radius.add(radius / 2)),
       ),
     );
   }
 
   static ShadCheckboxTheme checkboxTheme({
     required ShadColorScheme colorScheme,
+    required BorderRadius radius,
   }) {
     return ShadCheckboxTheme(
       size: 16,
       duration: 100.milliseconds,
-      radius: const BorderRadius.all(Radius.circular(4)),
       color: colorScheme.primary,
-      borderWidth: 1,
       padding: const EdgeInsets.only(left: 8),
       decoration: ShadDecoration(
-        border: const ShadBorder(width: 4, color: Colors.transparent),
-        focusedBorder: ShadBorder(
-          width: 2,
-          color: colorScheme.ring,
-          // the outer radius is calculated as
-          // outerRadius = innerRadius + innerRadius / 2
-          // outerRadius = 4 + 4 / 2 = 6
-          radius: const BorderRadius.all(Radius.circular(6)),
-          padding: const EdgeInsets.all(2),
+        border: ShadBorder(
+          width: 1,
+          color: colorScheme.primary,
+          radius: radius,
         ),
       ),
     );
@@ -420,51 +383,39 @@ abstract class ShadComponentDefaultTheme {
     required ShadTextTheme textTheme,
   }) {
     return ShadInputTheme(
+      style: textTheme.muted.copyWith(color: colorScheme.foreground),
+      placeholderStyle: textTheme.muted,
+      placeholderAlignment: Alignment.topLeft,
+      inputPadding: EdgeInsets.zero,
       decoration: ShadDecoration(
-        placeholderStyle: textTheme.muted,
-        placeholderAlignment: Alignment.topLeft,
         border: ShadBorder(
           width: 2,
-          color: Colors.transparent,
-          padding: const EdgeInsets.all(2),
+          color: colorScheme.border,
           radius: radius,
-        ),
-        focusedBorder: ShadBorder(
-          width: 2,
-          color: colorScheme.ring,
-          // the outer radius is calculated as
-          // outerRadius = innerRadius + innerRadius / 2
-          radius: radius.add(radius / 2),
-          padding: const EdgeInsets.all(2),
         ),
       ),
       padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-      border: Border.all(
-        color: colorScheme.border,
-        width: 2,
-      ),
-      radius: radius,
     );
   }
 
   static ShadRadioTheme radioTheme({
     required ShadColorScheme colorScheme,
   }) {
+    const circleSize = 10.0;
+    const radius = BorderRadius.all(Radius.circular(circleSize));
     return ShadRadioTheme(
       size: 16,
-      circleSize: 10,
+      circleSize: circleSize,
       duration: 100.milliseconds,
       color: colorScheme.primary,
-      borderWidth: 1,
       padding: const EdgeInsets.only(left: 8),
       decoration: ShadDecoration(
-        border: const ShadBorder(width: 4, color: Colors.transparent),
-        focusedBorder: ShadBorder(
-          width: 2,
-          color: colorScheme.ring,
-          radius: const BorderRadius.all(Radius.circular(16)),
-          padding: const EdgeInsets.all(2),
+        shape: BoxShape.circle,
+        border: ShadBorder(
+          width: 1,
+          color: colorScheme.primary,
         ),
+        secondaryFocusedBorder: ShadBorder(radius: radius.add(radius / 2)),
       ),
     );
   }
@@ -602,6 +553,8 @@ abstract class ShadComponentDefaultTheme {
       titleStyle: textTheme.large,
       descriptionStyle: textTheme.muted,
       alignment: Alignment.center,
+      mainAxisAlignment: MainAxisAlignment.start,
+      crossAxisAlignment: CrossAxisAlignment.stretch,
     );
   }
 
@@ -631,6 +584,8 @@ abstract class ShadComponentDefaultTheme {
       titleStyle: textTheme.large,
       descriptionStyle: textTheme.muted,
       alignment: Alignment.center,
+      mainAxisAlignment: MainAxisAlignment.start,
+      crossAxisAlignment: CrossAxisAlignment.stretch,
     );
   }
 
@@ -687,6 +642,7 @@ abstract class ShadComponentDefaultTheme {
       padding: const EdgeInsets.symmetric(vertical: 16),
       underlineTitleOnHover: true,
       duration: duration,
+      maintainState: false,
       iconEffects: const [
         RotateEffect(
           begin: 0,
@@ -696,23 +652,6 @@ abstract class ShadComponentDefaultTheme {
         ),
       ],
       curve: bezierCurve,
-      transitionBuilder: (animation, child) {
-        return AnimatedPadding(
-          padding: EdgeInsets.only(bottom: animation.value * 16),
-          curve: Curves.fastEaseInToSlowEaseOut,
-          duration: duration.divide(2),
-          child: SlideTransition(
-            position: Tween<Offset>(
-              begin: const Offset(0, 1),
-              end: Offset.zero,
-            ).animate(animation),
-            child: SizeTransition(
-              sizeFactor: animation,
-              child: child,
-            ),
-          ),
-        );
-      },
       titleStyle: textTheme.list.copyWith(
         fontWeight: FontWeight.w500,
       ),
@@ -746,11 +685,10 @@ abstract class ShadComponentDefaultTheme {
       dividerSize: 8,
       dividerColor: colorScheme.border,
       resetOnDoubleTap: true,
-      handleDecoration: BoxDecoration(
+      handleDecoration: ShadDecoration(
+        merge: false,
         color: colorScheme.border,
-        borderRadius: const BorderRadius.all(
-          Radius.circular(4),
-        ),
+        border: const ShadBorder(radius: BorderRadius.all(Radius.circular(4))),
       ),
       handleSize: const Size.square(10),
       mainAxisAlignment: MainAxisAlignment.start,
